Added cross-platform compiling(LINUX&Windows only)

+++ compile.py
@@ -1,27 +1,56 @@
 import os
+import sys
 
-os.system("rm -rf bin")
-os.system("mkdir bin")
+if (sys.platform == "linux"):
+    os.system("rm -rf bin")
+    os.system("mkdir bin")
 
-FilesToCompilePath = os.getcwd() + "/src"
+    FilesToCompilePath = os.getcwd() + "/src"
 
-os.chdir(FilesToCompilePath)
+    os.chdir(FilesToCompilePath)
 
-FilesInFolder = os.listdir()
+    FilesInFolder = os.listdir()
 
-FilesToCompile = []
+    FilesToCompile = []
 
-for files in FilesInFolder:
-    if files.find(".cpp") != -1:
-        FilesToCompile.append(files)
+    for files in FilesInFolder:
+        if files.find(".cpp") != -1 and files.find("WINDOWS") == -1:
+            FilesToCompile.append(files)
 
 
-command = "g++ -Wall -fdiagnostics-color=always -g"
+    command = "g++ -Wall -fdiagnostics-color=always -g"
 
-for files in FilesToCompile:
-    command += " "
-    command += files
+    for files in FilesToCompile:
+        command += " "
+        command += files
 
-command += " -o ../bin/VMC"
+    command += " -o ../bin/VMC"
 
-os.system(command)
+    os.system(command)
+
+if (sys.platform == "win32"):
+    os.system("rmdir /S /Q bin")
+    os.system("mkdir bin")
+
+    FilesToCompilePath = os.getcwd() + "\\src"
+
+    os.chdir(FilesToCompilePath)
+
+    FilesInFolder = os.listdir()
+
+    FilesToCompile = []
+
+    for files in FilesInFolder:
+        if files.find(".cpp") != -1 and files.find("LINUX") == -1:
+            FilesToCompile.append(files)
+
+
+    command = "g++ -Wall -fdiagnostics-color=always -g"
+
+    for files in FilesToCompile:
+        command += " "
+        command += files
+
+    command += " -o ..\\bin\\VMC"
+
+    os.system(command)
